Refactor into correct packages
