+++ src/main/java/seedu/address/logic/commands/events/RemindEventCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.events;
 
 import static java.util.Objects.requireNonNull;
 
@@ -8,13 +8,17 @@ import java.util.List;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.CommandResult;
+import seedu.address.logic.commands.CommandType;
+import seedu.address.logic.commands.CommandWord;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.event.Event;
 import seedu.address.model.event.Reminder;
 
 /**
- * Deletes an event identified using its displayed index from the Athena.
+ * Sets a custom reminder for an event identified using its displayed index from the Athena.
  */
 public class RemindEventCommand extends Command {
 
@@ -25,7 +29,7 @@ public class RemindEventCommand extends Command {
     public static final String MESSAGE_USAGE = COMMAND_WORD
             + ": Reminds you of the event \n"
             + "Parameters: INDEX (must be a positive integer)\n"
-            + "Example: " + COMMAND_WORD + " 1";
+            + "Example: " + COMMAND_WORD + " 1" + "in/10";
 
     public static final String MESSAGE_REMIND_EVENT_SUCCESS = "Reminder set for event %d days prior: ";
 
@@ -34,6 +38,10 @@ public class RemindEventCommand extends Command {
     private final Index targetIndex;
     private final int daysInAdvance;
 
+    /**
+     * @param targetIndex the index of the event that you want the reminder for
+     * @param daysInAdvance how many days in advance do you want to start getting reminders for this event
+     */
     public RemindEventCommand(Index targetIndex, int daysInAdvance) {
         this.targetIndex = targetIndex;
         this.daysInAdvance = daysInAdvance;
@@ -60,7 +68,8 @@ public class RemindEventCommand extends Command {
             throw new CommandException(Messages.MESSAGE_UNKNOWN_COMMAND);
         }
 
-        return new CommandResult(String.format(MESSAGE_REMIND_EVENT_SUCCESS, daysInAdvance) + eventForReminder.toString());
+        return new CommandResult(String.format(MESSAGE_REMIND_EVENT_SUCCESS, daysInAdvance)
+                + eventForReminder.toString());
     }
 
     @Override
