+++ src/main/java/seedu/address/logic/parser/events/RemindEventCommandParser.java
@@ -1,13 +1,14 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.events;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
-import static seedu.address.logic.parser.CliSyntax.*;
+import static seedu.address.logic.parser.CliSyntax.PREFIX_REMIND_IN;
 
 import java.util.stream.Stream;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
-import seedu.address.logic.commands.RemindEventCommand;
+import seedu.address.logic.commands.events.RemindEventCommand;
+import seedu.address.logic.parser.*;
 import seedu.address.logic.parser.exceptions.ParseException;
 
 
@@ -34,7 +35,8 @@ public class RemindEventCommandParser implements Parser<RemindEventCommand> {
         try {
             index = ParserUtil.parseIndex(argMultimap.getPreamble());
         } catch (IllegalValueException ive) {
-            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, RemindEventCommand.MESSAGE_USAGE), ive);
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT
+                    , RemindEventCommand.MESSAGE_USAGE), ive);
         }
 
         int daysInAdvance = Integer.parseInt(argMultimap.getValue(PREFIX_REMIND_IN).orElse("-1"));
